requirements.txt
django
tensorflow
numpy
opencv-python
Pillow
pywebview